Enum hypervisor::VcpuExit

pub enum VcpuExit {

    Io,
    Mmio,
    IoapicEoi {
        vector: u8,
    },
    HypervHypercall,
    Unknown,
    Exception,
    Hypercall,
    Debug,
    Hlt,
    IrqWindowOpen,
    Shutdown,
    FailEntry {
        hardware_entry_failure_reason: u64,
    },
    Intr,
    SetTpr,
    TprAccess,
    S390Sieic,
    S390Reset,
    Dcr,
    Nmi,
    InternalError,
    Osi,
    PaprHcall,
    S390Ucontrol,
    Watchdog,
    S390Tsch,
    Epr,
    SystemEventShutdown,
    SystemEventReset,
    SystemEventCrash,
    SystemEventS2Idle,
    RdMsr {
        index: u32,
    },
    WrMsr {
        index: u32,
        data: u64,
    },
    InvalidVpRegister,
    UnsupportedFeature,
    Canceled,
    UnrecoverableException,
    MsrAccess,
    Cpuid {
        entry: CpuIdEntry,
    },
    RdTsc,
    ApicSmiTrap,
    ApicInitSipiTrap,

}

A reason why a VCPU exited. One of these returns every time Vcpu::run is called.

Variants

Io

An io instruction needs to be emulated. vcpu handle_io should be called to handle the io operation

Mmio

A mmio instruction needs to be emulated. vcpu handle_mmio should be called to handle the mmio operation

IoapicEoi

Fields

vector: u8

HypervHypercall

Unknown

Exception

Hypercall

Debug

Hlt

IrqWindowOpen

Shutdown

FailEntry

Fields

hardware_entry_failure_reason: u64

Intr

SetTpr

TprAccess

S390Sieic

S390Reset

Dcr

Nmi

InternalError

Osi

PaprHcall

S390Ucontrol

Watchdog

S390Tsch

Epr

SystemEventShutdown

SystemEventReset

SystemEventCrash

SystemEventS2Idle

RdMsr

Fields

index: u32

WrMsr

Fields

index: u32

data: u64

InvalidVpRegister

An invalid vcpu register was set while running.

UnsupportedFeature

incorrect setup for vcpu requiring an unsupported feature

Canceled

vcpu run was user cancelled

UnrecoverableException

an unrecoverable exception was encountered (different from Exception)

MsrAccess

vcpu stopped due to an msr access.

Cpuid

Fields

entry: CpuIdEntry

vcpu stopped due to a cpuid request.

RdTsc

vcpu stopped due to calling rdtsc

ApicSmiTrap

vcpu stopped for an apic smi trap

ApicInitSipiTrap

vcpu stopped due to an apic trap