Struct hypervisor::kvm::KvmVcpu

pub struct KvmVcpu { /* fields omitted */ }

A wrapper around using a KVM Vcpu.

Implementations

impl KvmVcpu

pub fn pvclock_ctrl_arch(&self) -> Result<()>

Arch-specific implementation of Vcpu::pvclock_ctrl.

impl KvmVcpu

pub fn get_lapic(&self) -> Result<kvm_lapic_state>

X86 specific call to get the state of the "Local Advanced Programmable Interrupt Controller".

See the documentation for KVM_GET_LAPIC.

pub fn set_lapic(&self, klapic: &kvm_lapic_state) -> Result<()>

X86 specific call to set the state of the "Local Advanced Programmable Interrupt Controller".

See the documentation for KVM_SET_LAPIC.

impl KvmVcpu

pub fn get_mp_state(&self) -> Result<kvm_mp_state>

Gets the vcpu's current "multiprocessing state".

See the documentation for KVM_GET_MP_STATE. This call can only succeed after a call to Vm::create_irq_chip.

Note that KVM defines the call for both x86 and s390 but we do not expect anyone to run crosvm on s390.

pub fn set_mp_state(&self, state: &kvm_mp_state) -> Result<()>

Sets the vcpu's current "multiprocessing state".

See the documentation for KVM_SET_MP_STATE. This call can only succeed after a call to Vm::create_irq_chip.

Note that KVM defines the call for both x86 and s390 but we do not expect anyone to run crosvm on s390.