Enum hypervisor::VcpuExit

pub enum VcpuExit {

    IoOut {
        port: u16,
        size: usize,
        data: [u8; 8],
    },
    IoIn {
        port: u16,
        size: usize,
    },
    MmioRead {
        address: u64,
        size: usize,
    },
    MmioWrite {
        address: u64,
        size: usize,
        data: [u8; 8],
    },
    IoapicEoi {
        vector: u8,
    },
    HypervSynic {
        msr: u32,
        control: u64,
        evt_page: u64,
        msg_page: u64,
    },
    HypervHcall {
        input: u64,
        params: [u64; 2],
    },
    Unknown,
    Exception,
    Hypercall,
    Debug,
    Hlt,
    IrqWindowOpen,
    Shutdown,
    FailEntry {
        hardware_entry_failure_reason: u64,
    },
    Intr,
    SetTpr,
    TprAccess,
    S390Sieic,
    S390Reset,
    Dcr,
    Nmi,
    InternalError,
    Osi,
    PaprHcall,
    S390Ucontrol,
    Watchdog,
    S390Tsch,
    Epr,
    SystemEvent(u32, u64),

}

A reason why a VCPU exited. One of these returns every time Vcpu::run is called.

Variants

IoOut

An out port instruction was run on the given port with the given data.

Fields of IoOut

port: u16size: usizedata: [u8; 8]

IoIn

An in port instruction was run on the given port.

The data that the instruction receives should be set with set_data before Vcpu::run is called again.

Fields of IoIn

port: u16size: usize

MmioRead

A read instruction was run against the given MMIO address.

The data that the instruction receives should be set with set_data before Vcpu::run is called again.

Fields of MmioRead

address: u64size: usize

MmioWrite

A write instruction was run against the given MMIO address with the given data.

Fields of MmioWrite

address: u64size: usizedata: [u8; 8]

IoapicEoi

Fields of IoapicEoi

vector: u8

HypervSynic

Fields of HypervSynic

msr: u32control: u64evt_page: u64msg_page: u64

HypervHcall

Fields of HypervHcall

input: u64params: [u64; 2]

Unknown

Exception

Hypercall

Debug

Hlt

IrqWindowOpen

Shutdown

FailEntry

Fields of FailEntry

hardware_entry_failure_reason: u64

Intr

SetTpr

TprAccess

S390Sieic

S390Reset

Dcr

Nmi

InternalError

Osi

PaprHcall

S390Ucontrol

Watchdog

S390Tsch

Epr

SystemEvent(u32, u64)

The cpu triggered a system level event which is specified by the type field. The first field is the event type and the second field is flags. The possible event types are shutdown, reset, or crash. So far there are not any flags defined.

Tuple Fields of SystemEvent

0: u321: u64