Struct hypervisor::kvm::KvmVm

pub struct KvmVm { /* fields omitted */ }

A wrapper around creating and using a KVM VM.

Implementations

impl KvmVm

pub fn check_capability_arch(&self, c: VmCap) -> Option<bool>

Checks if a particular VmCap is available, or returns None if arch-independent Vm.check_capability() should handle the check.

pub fn get_device_params_arch(
&self,
_kind: DeviceKind
) -> Option<kvm_create_device>

Returns the params to pass to KVM_CREATE_DEVICE for a kind device on this arch, or None to let the arch-independent KvmVm::create_device handle it.

pub fn get_pvclock_arch(&self) -> Result<ClockState>

Arch-specific implementation of Vm::get_pvclock.

pub fn set_pvclock_arch(&self, state: &ClockState) -> Result<()>

Arch-specific implementation of Vm::set_pvclock.

pub fn get_pic_state(&self, id: PicSelect) -> Result<kvm_pic_state>

Retrieves the state of given interrupt controller by issuing KVM_GET_IRQCHIP ioctl.

Note that this call can only succeed after a call to Vm::create_irq_chip.

pub fn set_pic_state(&self, id: PicSelect, state: &kvm_pic_state) -> Result<()>

Sets the state of given interrupt controller by issuing KVM_SET_IRQCHIP ioctl.

Note that this call can only succeed after a call to Vm::create_irq_chip.

pub fn get_ioapic_num_pins(&self) -> Result<usize>

Retrieves the KVM_IOAPIC_NUM_PINS value for emulated IO-APIC.

pub fn get_ioapic_state(&self) -> Result<kvm_ioapic_state>

Retrieves the state of IOAPIC by issuing KVM_GET_IRQCHIP ioctl.

Note that this call can only succeed after a call to Vm::create_irq_chip.

pub fn set_ioapic_state(&self, state: &kvm_ioapic_state) -> Result<()>

Sets the state of IOAPIC by issuing KVM_SET_IRQCHIP ioctl.

Note that this call can only succeed after a call to Vm::create_irq_chip.

pub fn create_pit(&self) -> Result<()>

Creates a PIT as per the KVM_CREATE_PIT2 ioctl.

Note that this call can only succeed after a call to Vm::create_irq_chip.

pub fn get_pit_state(&self) -> Result<kvm_pit_state2>

Retrieves the state of PIT by issuing KVM_GET_PIT2 ioctl.

Note that this call can only succeed after a call to Vm::create_pit.

pub fn set_pit_state(&self, pit_state: &kvm_pit_state2) -> Result<()>

Sets the state of PIT by issuing KVM_SET_PIT2 ioctl.

Note that this call can only succeed after a call to Vm::create_pit.

pub fn enable_split_irqchip(&self, ioapic_pins: usize) -> Result<()>

Enable support for split-irqchip.

impl KvmVm

pub fn new(
kvm: &Kvm,
guest_mem: GuestMemory,
protection_type: ProtectionType
) -> Result<KvmVm>

Constructs a new KvmVm using the given Kvm instance.

pub fn create_irq_chip(&self) -> Result<()>

Creates an in kernel interrupt controller.

See the documentation on the KVM_CREATE_IRQCHIP ioctl.

pub fn set_irq_line(&self, irq: u32, active: bool) -> Result<()>

Sets the level on the given irq to 1 if active is true, and 0 otherwise.

pub fn register_irqfd(
&self,
gsi: u32,
evt: &Event,
resample_evt: Option<&Event>
) -> Result<()>

Registers an event that will, when signalled, trigger the gsi irq, and resample_evt ( when not None ) will be triggered when the irqchip is resampled.

pub fn unregister_irqfd(&self, gsi: u32, evt: &Event) -> Result<()>

Unregisters an event that was previously registered with register_irqfd.

The evt and gsi pair must be the same as the ones passed into register_irqfd.

pub fn set_gsi_routing(&self, routes: &[IrqRoute]) -> Result<()>

Sets the GSI routing table, replacing any table set with previous calls to set_gsi_routing.