Trait hypervisor::Vcpu

pub trait Vcpu: DowncastSync {
    fn try_clone(&self) -> Result<Self>
    where
        Self: Sized;

    fn as_vcpu(&self) -> &dyn Vcpu;

    fn take_run_handle(
        &self, 
        signal_num: Option<c_int>
    ) -> Result<VcpuRunHandle>;

    fn run(&self, run_handle: &VcpuRunHandle) -> Result<VcpuExit>;

    fn id(&self) -> usize;

    fn set_immediate_exit(&self, exit: bool);

    fn set_local_immediate_exit(exit: bool)
    where
        Self: Sized;

    fn set_local_immediate_exit_fn(&self) -> extern "C" fn();

    fn set_data(&self, data: &[u8]) -> Result<()>;

    fn pvclock_ctrl(&self) -> Result<()>;

    fn set_signal_mask(&self, signals: &[c_int]) -> Result<()>;

    unsafe fn enable_raw_capability(
        &self, 
        cap: u32, 
        args: &[u64; 4]
    ) -> Result<()>;
}

A virtual CPU holding a virtualized hardware thread's state, such as registers and interrupt state, which may be used to execute virtual machines.

To run, take_run_handle must be called to lock the vcpu to a thread. Then the returned VcpuRunHandle can be used for running.

Required methods

fn try_clone(&self) -> Result<Self> where
Self: Sized,

Makes a shallow clone of this Vcpu.

fn as_vcpu(&self) -> &dyn Vcpu

Casts this architecture specific trait object to the base trait object Vcpu.

fn take_run_handle(&self, signal_num: Option<c_int>) -> Result<VcpuRunHandle>

Returns a unique VcpuRunHandle. A VcpuRunHandle is required to run the guest.

Assigns a vcpu to the current thread so that signal handlers can call set_local_immediate_exit(). An optional signal number will be temporarily blocked while assigning the vcpu to the thread and later blocked when VcpuRunHandle is destroyed.

Returns an error, EBUSY, if the current thread already contains a Vcpu.

fn run(&self, run_handle: &VcpuRunHandle) -> Result<VcpuExit>

Runs the VCPU until it exits, returning the reason for the exit.

Note that the state of the VCPU and associated VM must be setup first for this to do anything useful. The given run_handle must be the same as the one returned by take_run_handle for this Vcpu.

fn id(&self) -> usize

Returns the vcpu id.

fn set_immediate_exit(&self, exit: bool)

Sets the bit that requests an immediate exit.

fn set_local_immediate_exit(exit: bool) where
Self: Sized,

Sets/clears the bit for immediate exit for the vcpu on the current thread.

fn set_local_immediate_exit_fn(&self) -> extern "C" fn()

Returns a function pointer that invokes set_local_immediate_exit in a signal-safe way when called.

fn set_data(&self, data: &[u8]) -> Result<()>

Sets the data received by a mmio read, ioport in, or hypercall instruction.

This function should be called after Vcpu::run returns an VcpuExit::IoIn, VcpuExit::MmioRead, or 'VcpuExit::HypervHcall`.

fn pvclock_ctrl(&self) -> Result<()>

Signals to the hypervisor that this guest is being paused by userspace. Only works on Vms that support VmCap::PvClockSuspend.

fn set_signal_mask(&self, signals: &[c_int]) -> Result<()>

Specifies set of signals that are blocked during execution of RunnableVcpu::run. Signals that are not blocked will cause run to return with VcpuExit::Intr. Only works on Vms that support VmCap::SignalMask.

unsafe fn enable_raw_capability(&self, cap: u32, args: &[u64; 4]) -> Result<()>

Enables a hypervisor-specific extension on this Vcpu. cap is a constant defined by the hypervisor API (e.g., kvm.h). args are the arguments for enabling the feature, if any.

Safety

This function is marked as unsafe because args may be interpreted as pointers for some capabilities. The caller must ensure that any pointers passed in the args array are allocated as the kernel expects, and that mutable pointers are owned.

Loading content...
Implementations

impl dyn Vcpu

pub fn is<__T: Vcpu>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Vcpu>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: Vcpu>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: Vcpu>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: Vcpu>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_arc<__T: Vcpu>(self: Arc<Self>) -> Result<Arc<__T>, Arc<Self>> where
__T: Any + Send + Sync,

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn't.