Trait hypervisor::x86_64::VcpuX86_64

pub trait VcpuX86_64: Vcpu {

    fn set_interrupt_window_requested(&self, requested: bool);

    fn ready_for_interrupt(&self) -> bool;

    fn interrupt(&self, irq: u32) -> Result<()>;

    fn inject_nmi(&self) -> Result<()>;

    fn get_regs(&self) -> Result<Regs>;

    fn set_regs(&self, regs: &Regs) -> Result<()>;

    fn get_sregs(&self) -> Result<Sregs>;

    fn set_sregs(&self, sregs: &Sregs) -> Result<()>;

    fn get_fpu(&self) -> Result<Fpu>;

    fn set_fpu(&self, fpu: &Fpu) -> Result<()>;

    fn get_debugregs(&self) -> Result<DebugRegs>;

    fn set_debugregs(&self, debugregs: &DebugRegs) -> Result<()>;

    fn get_xcrs(&self) -> Result<Vec<Register>>;

    fn set_xcrs(&self, xcrs: &[Register]) -> Result<()>;

    fn get_msrs(&self, msrs: &mut Vec<Register>) -> Result<()>;

    fn set_msrs(&self, msrs: &[Register]) -> Result<()>;

    fn set_cpuid(&self, cpuid: &CpuId) -> Result<()>;

    fn get_hyperv_cpuid(&self) -> Result<CpuId>;

    fn set_guest_debug(
        &self, 
        addrs: &[GuestAddress], 
        enable_singlestep: bool
    ) -> Result<()>;

    fn handle_cpuid(&mut self, entry: &CpuIdEntry) -> Result<()>;

    fn get_tsc_offset(&self) -> Result<u64>;

    fn set_tsc_offset(&self, offset: u64) -> Result<()>;

}

A wrapper around creating and using a VCPU on x86_64.

Required methods

fn set_interrupt_window_requested(&self, requested: bool)

Sets or clears the flag that requests the VCPU to exit when it becomes possible to inject interrupts into the guest.

fn ready_for_interrupt(&self) -> bool

Checks if we can inject an interrupt into the VCPU.

fn interrupt(&self, irq: u32) -> Result<()>

Injects interrupt vector irq into the VCPU.

fn inject_nmi(&self) -> Result<()>

Injects a non-maskable interrupt into the VCPU.

fn get_regs(&self) -> Result<Regs>

Gets the VCPU general purpose registers.

fn set_regs(&self, regs: &Regs) -> Result<()>

Sets the VCPU general purpose registers.

fn get_sregs(&self) -> Result<Sregs>

Gets the VCPU special registers.

fn set_sregs(&self, sregs: &Sregs) -> Result<()>

Sets the VCPU special registers.

fn get_fpu(&self) -> Result<Fpu>

Gets the VCPU FPU registers.

fn set_fpu(&self, fpu: &Fpu) -> Result<()>

Sets the VCPU FPU registers.

fn get_debugregs(&self) -> Result<DebugRegs>

Gets the VCPU debug registers.

fn set_debugregs(&self, debugregs: &DebugRegs) -> Result<()>

Sets the VCPU debug registers.

fn get_xcrs(&self) -> Result<Vec<Register>>

Gets the VCPU extended control registers.

fn set_xcrs(&self, xcrs: &[Register]) -> Result<()>

Sets the VCPU extended control registers.

fn get_msrs(&self, msrs: &mut Vec<Register>) -> Result<()>

Gets the model-specific registers. msrs specifies the MSR indexes to be queried, and on success contains their indexes and values.

fn set_msrs(&self, msrs: &[Register]) -> Result<()>

Sets the model-specific registers.

fn set_cpuid(&self, cpuid: &CpuId) -> Result<()>

Sets up the data returned by the CPUID instruction.

fn get_hyperv_cpuid(&self) -> Result<CpuId>

Gets the system emulated hyper-v CPUID values.

fn set_guest_debug(
&self,
addrs: &[GuestAddress],
enable_singlestep: bool
) -> Result<()>

Sets up debug registers and configure vcpu for handling guest debug events.

fn handle_cpuid(&mut self, entry: &CpuIdEntry) -> Result<()>

This function should be called after Vcpu::run returns VcpuExit::Cpuid, and entry should represent the result of emulating the CPUID instruction. The handle_cpuid function will then set the appropriate registers on the vcpu.

fn get_tsc_offset(&self) -> Result<u64>

Get the guest->host TSC offset

fn set_tsc_offset(&self, offset: u64) -> Result<()>

Set the guest->host TSC offset

Implementations

impl dyn VcpuX86_64

pub fn is<__T: VcpuX86_64>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: VcpuX86_64>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: VcpuX86_64>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: VcpuX86_64>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: VcpuX86_64>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.